+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 ## Changelog
 
+### Version 6.0.0-alpha.7
+- All: clean JS warnings (https://github.com/react-native-video/react-native-video/pull/3183)
+- Android: Add shutterView color configurtion (https://github.com/react-native-video/react-native-video/pull/3179)
+- Android: React native 0.73 support (https://github.com/react-native-video/react-native-video/pull/3163)
+- Android: Fix memory leaks from AudioManager [#3123](https://github.com/react-native-video/react-native-video/pull/3123)
+- Android: Fixed syntax error [#3182](https://github.com/react-native-video/react-native-video/issues/3182)
+- iOS: Fix freeze at playback startup (https://github.com/react-native-video/react-native-video/pull/3173)
+- iOS: Various safety checks (https://github.com/react-native-video/react-native-video/pull/3168)
+
+### Version 6.0.0-alpha.6
+- Feature: Video range support [#3030](https://github.com/react-native-video/react-native-video/pull/3030)
+- iOS: remove undocumented `currentTime` property [#3064](https://github.com/react-native-video/react-native-video/pull/3064)
+- iOS: make sure that the audio in ads is muted when the player is muted. [#3068](https://github.com/react-native-video/react-native-video/pull/3077)
+- iOS: make IMA build optionnal
+
 ### Version 6.0.0-alpha.5
 
 - iOS: ensure controls are not displayed when disabled by user [#3017](https://github.com/react-native-video/react-native-video/pull/3017)